feat(setup): replace language seeding with setup in navigation
We provide seed demo data fr german (switzerland) and partially
for english and the netherlands. It is dangerous if these seed
files are re-loaded into existing data.

There exists a preliminary setup wizard in the hide.

This patch replaces the menu items to load the seed data with
a link to the setup wizard.

The wizard currently shows two page-forms to setup the tenant.