fix(Bank Account): fix creation of bank accounts
We have common accounts and bank accounts. The bank accounts are
assigned to a bank. There is no intuitive way of creating a bank
account. You need to edit or view a bank account, go to the list and
then create a new one. For this reason it took a long time to trigger a
bug in this process.

This patch fixes the creation of a new bank account by configuring the
bank account controller to use BankAccount as its model.