fix: make application.yml optional
We do support application level settings for application title and other
things like API keys for Google analytics. These are configured in
config/application.yml. We also provide an example version for this
file. The .example version should be used if no actual application.yml
exists.

This patch drops the config/application.yml and adds it to .gitignore.
The Settings class is configured to use the
config/application.yml.example as a fallback.