feat(tenant): we only support one tenant
When we started this project the idea was to support multiple tenants.
This is now a reality. But when we actually implemented the multi
tenancy we used multiple databases for that and introduced the
Admin::Tenant model. The Tenant model is actually mostly a holder of
per tenant 'settings'.

While some code can already handle only one Tenant instance we had some
UI to create or even delete Tenant instances. We got even a fix for the
creation of new Tenants.

This patch drops the UI for Tenant CRUD operations.