feat(capistrano): re-add capistrano for deployment
We used capistrano with our own capones-recipes extensions to deploy
Bookyt. Capones-recipes allows to hold deployment stages in another
repository, a feature we (CyT) uses actively. It also provides some nice
helpers to sync all data from a deployment down and so on.

The capones-recipes gem did not get much love though, and is therefore a
bit outdated and misses docs. For that reason it was dropped in the last
round of gem updates. We need to do deployments though, and therefore
add it back for now.

This commit add all the files back that are needed to do capistrano
deployments beside actual deployment specifications.