feat: implement global fiscal year filter state
When doing accounting, one focuses mostly on a single fiscal year. We
use that information to filter in a couple of view what we show. We did
already implement such a filter, but it was stateless, only applying to
the current view.

This patch introduces a global fiscal year state.

Dropping of the sidebar filters come in a follow up patch.