refactor(user): drop unused User.current_company
We used to determine the company from the current user and did that by
going through the User.person.employment association. But not each
Person object assigned to User has an #employment association. And we do
not need this way of company lookup anymore.

This patch drops the User.current_company method plus associated test.