fix(migrations): drop password_salt column in all tenants
There's a database cleanup migration that was dropping a column on a
table if present. But it did not reload the column definitions before
doing that. In the case of Apartment multi tenants this means the first
Tenant (public) was actually telling if the column should be dropped or
not.