Switch from tags to codes to reference accounts
Unfortunately we cannot use tags to reference accounts. When returning
the line items of an invoice, we cannot know the tag which has been used
to reference the accounts and if its unique.
This prevents updates to an invoice or bookings as the update call expects
tags for referencing.

As codes are unique and also more intutive than the id of an account,
it's the way better choice to use than tags.